+++ RELEASE-NOTES.md
@@ -1,6 +1,13 @@
 ddth-dao release notes
 ======================
 
+0.5.1.2 - 2016-04-22
+--------------------
+
+- Minor enhancements: no more `synchronized` methods in class `BaseBo`.
+- `BaseBo`: New method `protected void triggerPopulate()`
+
+
 0.5.0.1 - 2015-10-09
 --------------------